requirements.txt
python-dotenv
requests
Flask==3.1.2
